requirements.txt
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
db-dtypes>=1.1.1

